+++ pyndows/testing.py
@@ -6,10 +6,16 @@ import datetime
 
 import pytest
 
-from smb.smb_structs import OperationFailure
+from smb.smb_structs import (
+    OperationFailure,
+    SMB_FILE_ATTRIBUTE_DIRECTORY,
+    SMB_FILE_ATTRIBUTE_READONLY,
+    SMB_FILE_ATTRIBUTE_ARCHIVE,
+    SMB_FILE_ATTRIBUTE_INCL_NORMAL,
+)
 from smb.base import SharedFile
 
-SharedFileMock = namedtuple("SharedFileMock", ["filename"])
+SharedFileMock = namedtuple("SharedFileMock", ["filename", "isDirectory"])
 
 
 class _FileStore(dict):
@@ -115,16 +121,35 @@ class SMBConnectionMock:
         raise OperationFailure("Mock for retrieveFile failure.", [])
 
     def listPath(
-        self, service_name: str, path: str, pattern: str = "*"
+        self, service_name: str, path: str, search: int = 0, pattern: str = "*",
     ) -> List[SharedFile]:
-        files_list = [
-            SharedFileMock(os.path.basename(file_path))
-            for _, file_path in SMBConnectionMock.stored_files
-            if re.search(pattern, os.path.basename(file_path))
-        ]
+        pattern = pattern.replace("*", ".*")
+        service_name = service_name.replace("/", os.sep).replace("\\", os.sep)
+        path = path.replace("/", os.sep).replace("\\", os.sep)
+        path = path if not path or path.endswith(os.sep) else f"{path}{os.sep}"
+
+        def to_file_in_path(file_path: str, path: str) -> (str, bool):
+            if f"{os.path.dirname(file_path)}{os.sep}".startswith(path):
+                file_path = file_path[len(path) :]
+                return file_path.split(os.sep, maxsplit=1)[0], os.sep in file_path
+            return None, None
+
+        files_list = []
+        for shared_folder, file_path in SMBConnectionMock.stored_files:
+            shared_folder = shared_folder.replace("/", os.sep).replace("\\", os.sep)
+            file_path = file_path.replace("/", os.sep).replace("\\", os.sep)
+            file_name, is_directory = to_file_in_path(file_path, path)
+            if (
+                file_name
+                and shared_folder == service_name
+                and re.search(pattern, file_name)
+            ):
+                if search | SMB_FILE_ATTRIBUTE_DIRECTORY == search or not is_directory:
+                    files_list.append(SharedFileMock(file_name, is_directory))
+
         if not files_list:
             raise OperationFailure("Mock for listPath failure.", [])
-        return files_list
+        return sorted(set(files_list), key=lambda file: file.filename)
 
     def echo(self, data, timeout: int = 10):
         echo_response = SMBConnectionMock.echo_responses.pop(data, None)