+++ tests/test_windows.py
@@ -1,11 +1,10 @@
 import os
 import os.path
-import tempfile
 
 import pytest
 
 import pyndows
-from pyndows.testing import samba_mock, SMBConnectionMock
+from pyndows.testing import samba_mock, SMBConnectionMock, SharedFileMock
 
 
 def test_successful_connection(samba_mock: SMBConnectionMock):
@@ -41,106 +40,87 @@ def test_connection_timeout(samba_mock: SMBConnectionMock):
     )
 
 
-def test_file_retrieval(samba_mock: SMBConnectionMock):
+def test_file_retrieval(samba_mock: SMBConnectionMock, tmpdir):
     connection = pyndows.connect(
         "TestComputer", "127.0.0.1", 80, "TestDomain", "TestUser", "TestPassword"
     )
-    with tempfile.TemporaryDirectory() as temp_dir:
-        samba_mock.files_to_retrieve[("TestShare", "TestFilePath")] = "Test Content"
+    samba_mock.files_to_retrieve[("TestShare", "TestFilePath")] = "Test Content"
 
+    pyndows.get(
+        connection, "TestShare", "TestFilePath", os.path.join(tmpdir, "local_file"),
+    )
+    with open(os.path.join(tmpdir, "local_file")) as local_file:
+        assert local_file.read() == "Test Content"
+
+
+def test_operation_failure_during_file_retrieval(samba_mock: SMBConnectionMock, tmpdir):
+    connection = pyndows.connect(
+        "TestComputer", "127.0.0.1", 80, "TestDomain", "TestUser", "TestPassword"
+    )
+    with pytest.raises(Exception) as exception_info:
         pyndows.get(
-            connection,
-            "TestShare",
-            "TestFilePath",
-            os.path.join(temp_dir, "local_file"),
+            connection, "TestShare", "TestFilePath", os.path.join(tmpdir, "local_file"),
         )
-        with open(os.path.join(temp_dir, "local_file")) as local_file:
-            assert local_file.read() == "Test Content"
+    assert (
+        str(exception_info.value)
+        == r"Unable to retrieve \\TestComputer\TestShareTestFilePath file"
+    )
 
 
-def test_operation_failure_during_file_retrieval(samba_mock: SMBConnectionMock):
+def test_file_move(samba_mock: SMBConnectionMock, tmpdir):
     connection = pyndows.connect(
         "TestComputer", "127.0.0.1", 80, "TestDomain", "TestUser", "TestPassword"
     )
-    with tempfile.TemporaryDirectory() as temp_dir:
-        with pytest.raises(Exception) as exception_info:
-            pyndows.get(
-                connection,
-                "TestShare",
-                "TestFilePath",
-                os.path.join(temp_dir, "local_file"),
-            )
-        assert (
-            str(exception_info.value)
-            == r"Unable to retrieve \\TestComputer\TestShareTestFilePath file"
-        )
+    with open(os.path.join(tmpdir, "local_file"), mode="w") as distant_file:
+        distant_file.write("Test Content Move")
+
+    pyndows.move(
+        connection, "TestShare", "TestFilePath", os.path.join(tmpdir, "local_file"),
+    )
+
+    assert samba_mock.stored_files[("TestShare", "TestFilePath")] == "Test Content Move"
 
 
-def test_file_move(samba_mock: SMBConnectionMock):
+def test_store_file_operation_failure_during_file_move(
+    samba_mock: SMBConnectionMock, tmpdir
+):
     connection = pyndows.connect(
         "TestComputer", "127.0.0.1", 80, "TestDomain", "TestUser", "TestPassword"
     )
-    with tempfile.TemporaryDirectory() as temp_dir:
-        with open(os.path.join(temp_dir, "local_file"), mode="w") as distant_file:
-            distant_file.write("Test Content Move")
+    with open(os.path.join(tmpdir, "local_file"), mode="w") as distant_file:
+        distant_file.write("Test Content Move")
 
+    samba_mock.storeFile_failure = True
+    with pytest.raises(Exception) as exception_info:
         pyndows.move(
-            connection,
-            "TestShare",
-            "TestFilePath",
-            os.path.join(temp_dir, "local_file"),
+            connection, "TestShare", "TestFilePath", os.path.join(tmpdir, "local_file"),
         )
 
-        assert (
-            samba_mock.stored_files[("TestShare", "TestFilePath")]
-            == "Test Content Move"
-        )
+    assert (
+        str(exception_info.value)
+        == r"Unable to write \\TestComputer\TestShareTestFilePath.tmp"
+    )
 
 
-def test_store_file_operation_failure_during_file_move(samba_mock: SMBConnectionMock):
+def test_rename_operation_failure_during_file_move(
+    samba_mock: SMBConnectionMock, tmpdir
+):
     connection = pyndows.connect(
         "TestComputer", "127.0.0.1", 80, "TestDomain", "TestUser", "TestPassword"
     )
-    with tempfile.TemporaryDirectory() as temp_dir:
-        with open(os.path.join(temp_dir, "local_file"), mode="w") as distant_file:
-            distant_file.write("Test Content Move")
+    with open(os.path.join(tmpdir, "local_file"), mode="w") as distant_file:
+        distant_file.write("Test Content Move")
 
-        samba_mock.storeFile_failure = True
-        with pytest.raises(Exception) as exception_info:
-            pyndows.move(
-                connection,
-                "TestShare",
-                "TestFilePath",
-                os.path.join(temp_dir, "local_file"),
-            )
-
-        assert (
-            str(exception_info.value)
-            == r"Unable to write \\TestComputer\TestShareTestFilePath.tmp"
+    samba_mock.rename_failure = True
+    with pytest.raises(Exception) as exception_info:
+        pyndows.move(
+            connection, "TestShare", "TestFilePath", os.path.join(tmpdir, "local_file"),
         )
 
-
-def test_rename_operation_failure_during_file_move(samba_mock: SMBConnectionMock):
-    connection = pyndows.connect(
-        "TestComputer", "127.0.0.1", 80, "TestDomain", "TestUser", "TestPassword"
+    assert (
+        str(exception_info.value)
+        == r"Unable to rename temp file into \\TestComputer\TestShareTestFilePath"
     )
-    with tempfile.TemporaryDirectory() as temp_dir:
-        with open(os.path.join(temp_dir, "local_file"), mode="w") as distant_file:
-            distant_file.write("Test Content Move")
-
-        samba_mock.rename_failure = True
-        with pytest.raises(Exception) as exception_info:
-            pyndows.move(
-                connection,
-                "TestShare",
-                "TestFilePath",
-                os.path.join(temp_dir, "local_file"),
-            )
-
-        assert (
-            str(exception_info.value)
-            == r"Unable to rename temp file into \\TestComputer\TestShareTestFilePath"
-        )
 
 
 def test_file_rename(samba_mock: SMBConnectionMock):
@@ -185,6 +165,270 @@ def test_file_rename_file_does_not_exist(samba_mock: SMBConnectionMock):
         str(exception_info.value)
         == r"\\TestComputer\TestShare\file_to_rename_2 doesn't exist"
     )
+
+
+def test_get_all_folder_contents(samba_mock: SMBConnectionMock):
+    connection = pyndows.connect(
+        "TestComputer", "127.0.0.1", 80, "TestDomain", "TestUser", "TestPassword"
+    )
+
+    samba_mock.stored_files.update(
+        {
+            ("TestShare/", "1"): "Test Find",
+            ("TestShare/", "A/1"): "Test Find",
+            ("TestShare/", "A/2"): "Test Find",
+            ("TestShare/", "A/3"): "Test Find",
+            ("TestShare/", "A/i/1"): "Test Find",
+            ("TestShare/", "A/i/2"): "Test Find",
+        }
+    )
+
+    shared_folder_contents = pyndows.get_folder_content(
+        connection, "TestShare/", folder_path="A"
+    )
+
+    assert shared_folder_contents == [
+        SharedFileMock(filename="1", isDirectory=False),
+        SharedFileMock(filename="2", isDirectory=False),
+        SharedFileMock(filename="3", isDirectory=False),
+        SharedFileMock(filename="i", isDirectory=True),
+    ]
+
+
+def test_get_all_folder_contents_providing_paths_with_backslashes(
+    samba_mock: SMBConnectionMock,
+):
+    connection = pyndows.connect(
+        "TestComputer", "127.0.0.1", 80, "TestDomain", "TestUser", "TestPassword"
+    )
+
+    samba_mock.stored_files.update(
+        {
+            ("TestShare\\", "1"): "Test Find",
+            ("TestShare\\", "A\\1"): "Test Find",
+            ("TestShare\\", "A\\2"): "Test Find",
+            ("TestShare\\", "A\\3"): "Test Find",
+            ("TestShare\\", "A\\i\\1"): "Test Find",
+            ("TestShare\\", "A\\i\\2"): "Test Find",
+        }
+    )
+
+    shared_folder_contents = pyndows.get_folder_content(
+        connection, "TestShare\\", folder_path="A"
+    )
+
+    assert shared_folder_contents == [
+        SharedFileMock(filename="1", isDirectory=False),
+        SharedFileMock(filename="2", isDirectory=False),
+        SharedFileMock(filename="3", isDirectory=False),
+        SharedFileMock(filename="i", isDirectory=True),
+    ]
+
+
+def test_get_all_sub_folder_contents_providing_paths_with_a_mix_of_slashes_and_backslashes(
+    samba_mock: SMBConnectionMock,
+):
+    connection = pyndows.connect(
+        "TestComputer", "127.0.0.1", 80, "TestDomain", "TestUser", "TestPassword"
+    )
+
+    samba_mock.stored_files.update(
+        {
+            ("TestShare/", "1"): "Test Find",
+            ("TestShare/", "A/1"): "Test Find",
+            ("TestShare\\", "A/2"): "Test Find",
+            ("TestShare/", "A/3"): "Test Find",
+            ("TestShare/", "A/i\\1"): "Test Find",
+            ("TestShare\\", "A/i/2"): "Test Find",
+        }
+    )
+
+    shared_folder_contents = pyndows.get_folder_content(
+        connection, "TestShare/", folder_path="A\\i"
+    )
+
+    assert shared_folder_contents == [
+        SharedFileMock(filename="1", isDirectory=False),
+        SharedFileMock(filename="2", isDirectory=False),
+    ]
+
+
+def test_get_all_folder_contents_empty_folder(samba_mock: SMBConnectionMock):
+    connection = pyndows.connect(
+        "TestComputer", "127.0.0.1", 80, "TestDomain", "TestUser", "TestPassword"
+    )
+
+    samba_mock.stored_files[("TestShare/", "1")] = "Test Find"
+
+    shared_folder_contents = pyndows.get_folder_content(
+        connection, "TestShare/", folder_path="A"
+    )
+
+    assert shared_folder_contents == []
+
+
+def test_get_all_folder_contents_empty_shared_folder(samba_mock: SMBConnectionMock):
+    connection = pyndows.connect(
+        "TestComputer", "127.0.0.1", 80, "TestDomain", "TestUser", "TestPassword"
+    )
+
+    shared_folder_contents = pyndows.get_folder_content(
+        connection, "TestShare/", folder_path=""
+    )
+
+    assert shared_folder_contents == []
+
+
+def test_get_all_folder_contents_non_existing_folder(samba_mock: SMBConnectionMock):
+    connection = pyndows.connect(
+        "TestComputer", "127.0.0.1", 80, "TestDomain", "TestUser", "TestPassword"
+    )
+
+    shared_folder_contents = pyndows.get_folder_content(
+        connection, "TestShare/", folder_path="B"
+    )
+
+    assert shared_folder_contents == []
+
+
+def test_get_all_folder_contents_non_existing_shared_folder(
+    samba_mock: SMBConnectionMock,
+):
+    connection = pyndows.connect(
+        "TestComputer", "127.0.0.1", 80, "TestDomain", "TestUser", "TestPassword"
+    )
+
+    samba_mock.stored_files[("TestShare/", "1")] = "Test Find"
+
+    shared_folder_contents = pyndows.get_folder_content(
+        connection, "AnotherTestShare/", folder_path=""
+    )
+
+    assert shared_folder_contents == []
+
+
+def test_get_all_folder_removes_self_directory_and_parent_directory(
+    samba_mock: SMBConnectionMock,
+):
+    connection = pyndows.connect(
+        "TestComputer", "127.0.0.1", 80, "TestDomain", "TestUser", "TestPassword"
+    )
+
+    samba_mock.stored_files.update(
+        {("TestShare/", "."): "Test Find", ("TestShare/", ".."): "Test Find",}
+    )
+
+    shared_folder_contents = pyndows.get_folder_content(
+        connection, "TestShare/", folder_path=""
+    )
+
+    assert shared_folder_contents == []
+
+
+def test_get_all_folder_contents_excluding_directories(samba_mock: SMBConnectionMock):
+    connection = pyndows.connect(
+        "TestComputer", "127.0.0.1", 80, "TestDomain", "TestUser", "TestPassword"
+    )
+
+    samba_mock.stored_files.update(
+        {
+            ("TestShare/", "1"): "Test Find",
+            ("TestShare/", "A/1"): "Test Find",
+            ("TestShare/", "A/2"): "Test Find",
+            ("TestShare/", "A/3"): "Test Find",
+            ("TestShare/", "A/i/1"): "Test Find",
+            ("TestShare/", "A/i/2"): "Test Find",
+        }
+    )
+
+    shared_folder_contents = pyndows.get_folder_content(
+        connection, "TestShare/", folder_path="A", include_folders=False
+    )
+
+    assert shared_folder_contents == [
+        SharedFileMock(filename="1", isDirectory=False),
+        SharedFileMock(filename="2", isDirectory=False),
+        SharedFileMock(filename="3", isDirectory=False),
+    ]
+
+
+def test_get_all_folder_contents_matching_a_pattern(samba_mock: SMBConnectionMock):
+    connection = pyndows.connect(
+        "TestComputer", "127.0.0.1", 80, "TestDomain", "TestUser", "TestPassword"
+    )
+
+    samba_mock.stored_files.update(
+        {
+            ("TestShare/", "1"): "Test Find",
+            ("TestShare/", "A/1"): "Test Find",
+            ("TestShare/", "A/12"): "Test Find",
+            ("TestShare/", "A/3"): "Test Find",
+            ("TestShare/", "A/i/1"): "Test Find",
+            ("TestShare/", "A/1i/2"): "Test Find",
+        }
+    )
+
+    shared_folder_contents = pyndows.get_folder_content(
+        connection, "TestShare/", folder_path="A", pattern="^1"
+    )
+
+    assert shared_folder_contents == [
+        SharedFileMock(filename="1", isDirectory=False),
+        SharedFileMock(filename="12", isDirectory=False),
+        SharedFileMock(filename="1i", isDirectory=True),
+    ]
+
+
+def test_get_all_shared_folder_contents(samba_mock: SMBConnectionMock):
+    connection = pyndows.connect(
+        "TestComputer", "127.0.0.1", 80, "TestDomain", "TestUser", "TestPassword"
+    )
+
+    samba_mock.stored_files.update(
+        {
+            ("TestShare/", "1"): "Test Find",
+            ("TestShare/", "A/1"): "Test Find",
+            ("TestShare/", "A/2"): "Test Find",
+            ("TestShare/", "A/3"): "Test Find",
+            ("TestShare/", "A/i/1"): "Test Find",
+            ("TestShare/", "A/i/2"): "Test Find",
+        }
+    )
+
+    shared_folder_contents = pyndows.get_folder_content(
+        connection, "TestShare/", folder_path=""
+    )
+
+    assert shared_folder_contents == [
+        SharedFileMock(filename="1", isDirectory=False),
+        SharedFileMock(filename="A", isDirectory=True),
+    ]
+
+
+def test_get_all_sub_folder_contents(samba_mock: SMBConnectionMock):
+    connection = pyndows.connect(
+        "TestComputer", "127.0.0.1", 80, "TestDomain", "TestUser", "TestPassword"
+    )
+
+    samba_mock.stored_files.update(
+        {
+            ("TestShare/", "1"): "Test Find",
+            ("TestShare/", "A/1"): "Test Find",
+            ("TestShare/", "A/2"): "Test Find",
+            ("TestShare/", "A/3"): "Test Find",
+            ("TestShare/", "A/i/1"): "Test Find",
+            ("TestShare/", "A/i/2"): "Test Find",
+        }
+    )
+
+    shared_folder_contents = pyndows.get_folder_content(
+        connection, "TestShare/", folder_path="A/i"
+    )
+
+    assert shared_folder_contents == [
+        SharedFileMock(filename="1", isDirectory=False),
+        SharedFileMock(filename="2", isDirectory=False),
+    ]
 
 
 def test_get_file_desc_file_exists(samba_mock: SMBConnectionMock):